Refactor codebase, move parsers/types to separate files
