+++ rust-app/src/interface.rs
@@ -1,7 +1,7 @@
+use crate::parser::common::*;
 use core::convert::TryFrom;
 use core::marker::ConstParamTy;
 use ledger_device_sdk::io::{ApduHeader, StatusWords};
-use ledger_parser_combinators::bcs::async_parser::*;
 use ledger_parser_combinators::core_parsers::*;
 use ledger_parser_combinators::endianness::*;
 use num_enum::TryFromPrimitive;
@@ -16,74 +16,6 @@ pub enum ParseChecks {
 
 // Payload for a public key request
 pub type Bip32Key = DArray<Byte, U32<{ Endianness::Little }>, 10>;
-
-pub type SignParameters = (IntentMessage, Bip32Key);
-
-// Sui Types
-pub type IntentMessage = (Intent, TransactionData);
-
-pub struct TransactionData;
-
-pub type TransactionDataV1 = (
-    TransactionKind,
-    SuiAddress,            // sender
-    GasData,               // gas_data
-    TransactionExpiration, // expiration
-);
-
-pub struct TransactionKind;
-
-pub struct ProgrammableTransaction;
-
-pub struct CommandSchema;
-pub struct ArgumentSchema;
-pub struct CallArgSchema;
-
-pub type GasData = (
-    Vec<ObjectRef, { usize::MAX }>, // payment
-    SuiAddress,                     // owner
-    Amount,                         // price
-    Amount,                         // budget
-);
-
-pub struct TransactionExpiration;
-pub type EpochId = U64<{ Endianness::Little }>;
-
-pub type ObjectRef = (ObjectID, SequenceNumber, ObjectDigest);
-
-pub type SharedObject = (
-    ObjectID,       // id
-    SequenceNumber, // initial_shared_version
-    bool,           // mutable
-);
-
-pub type AccountAddress = SuiAddress;
-pub type ObjectID = AccountAddress;
-pub type SequenceNumber = U64LE;
-pub type ObjectDigest = SHA3_256_HASH;
-
-pub const SUI_ADDRESS_LENGTH: usize = 32;
-pub type SuiAddress = Array<Byte, SUI_ADDRESS_LENGTH>;
-
-pub type Coins = Vec<ObjectRef, { usize::MAX }>;
-
-pub type Recipient = SuiAddress;
-
-pub type Amount = U64LE;
-
-pub type U64LE = U64<{ Endianness::Little }>;
-pub type U16LE = U16<{ Endianness::Little }>;
-
-pub type Intent = (IntentVersion, IntentScope, AppId);
-pub type IntentVersion = ULEB128;
-pub type IntentScope = ULEB128;
-pub type AppId = ULEB128;
-
-// TODO: confirm if 33 is indeed ok for all uses of SHA3_256_HASH
-#[allow(non_camel_case_types)]
-pub type SHA3_256_HASH = Array<Byte, 33>;
-
-pub type SuiAddressRaw = [u8; SUI_ADDRESS_LENGTH];
 
 #[allow(dead_code)]
 pub struct SuiPubKeyAddress(ledger_device_sdk::ecc::ECPublicKey<65, 'E'>, SuiAddressRaw);
