+++ rust-app/src/swap/params.rs
@@ -4,7 +4,7 @@ use core::ffi::CStr;
 use core::mem;
 use ledger_device_sdk::libcall;
 
-use crate::interface::SuiAddressRaw;
+use crate::parser::common::SuiAddressRaw;
 use crate::swap::Error;
 
 // Max SUI address str length is 32*2
